Rename gen cmd to gen-types
  - Also make room for future commands to be added.
